feature: 连接到直播间

+++ chatbilibili-api/src/main/java/top/nino/api/model/room/RoomInit.java
@@ -5,17 +5,20 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 
+/**
+ * @author nino
+ */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class RoomInit {
 	private Long room_id;
-	private Integer short_id;
+	private Long short_id;
 	private Long uid;
 	private Short need_p2p;
 	private Boolean is_hidden;
 	private Boolean is_portrait;
-	//0为不直播  1为直播中 2为轮播
+	// 0为不直播  1为直播中 2为轮播
 	private Integer live_status;
 	private Short hidden_till;
 	private Short lock_till;
@@ -26,5 +29,5 @@ public class RoomInit {
 	private Short is_sp;
 	private Short special_type;
 
-	
+
 }
